+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add Fabric API provisioning support
+
 ## 1.0.0 - 2025-11-29
 ### Added
 - Add fabmc_server_properties configuration for setting server.properties values